requirements.txt
python-nmap
git+ssh://git@github.com/stealthizer/anthill.git@master#master
